seperated server grpc service from actually running the server for easier testing
